+++ editor/grida-canvas/hotkey/index.ts
